requirements.txt
sentence-transformers 
chromadb 
google-generativeai 
gradio